% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJxCounts.R
\name{getJxCounts}
\alias{getJxCounts}
\title{Obtain Exonic Feature Counts}
\usage{
getJxCounts(input, local = F)
}
\arguments{
\item{input}{Either a SRA project name (part of recount2) OR
the path to the rail output file containing junction information of the form 'cross_sample_results/junctions.tsv.gz'.}
}
\description{
This function returns the coverage of the junction portion of the sufficient
features identified based on sequencing read length.
}
\keyword{getJxCounts}
