% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPheno.R
\name{processPheno}
\alias{processPheno}
\title{Phenotype Information Creation}
\usage{
processPheno(input)
}
\arguments{
\item{input}{The project name from SRA as a \code{character}
OR a \code{data.frame} describing another project aligned with Rail-RNA.}
}
\value{
A phenotype matrix containing at least the necessary information to run recountNNLS.
}
\description{
This function creates a phenotype/meta-information matrix with the information needed
for the rest of the analysis. The supplied input is either an SRA
project id (from a project compiled on recount2), or a manifest file
that contains information where the Rail-RNA outputs are.
}
\examples{
project = 'SRP063581'
pheno = processPheno(project)
}
\keyword{processPheno}
