% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processReadLength.R
\name{processReadLength}
\alias{processReadLength}
\title{NNLS for Tx Abundance Calculation}
\usage{
processReadLength(rl, pheno, counts_ex, counts_jx, cores)
}
\arguments{
\item{rl}{A read length that exists in the pheno table}

\item{pheno}{The phenotype matrix created by processPheno().}

\item{counts_ex}{The matrix of exonic feature counts where each row represents a exonic feature and each column a sample.}

\item{counts_jx}{The matrix of junction feature counts where each row represents a junction feature and each column a sample.}

\item{cores}{The number of processing cores to use.}
}
\description{
This function applies runs our NNLS model for a given subset
of the pheno matrix that has the same read length specified
and reports an rse.
}
\keyword{recountNNLS}
