% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExCounts.R
\name{getExCounts}
\alias{getExCounts}
\title{Obtain Exonic Feature Counts}
\usage{
getExCounts(pheno, cores = 1)
}
\arguments{
\item{pheno}{The phenotype matrix created by processPheno().}
}
\description{
This function returns the coverage of the exonic portion of sufficient
features identified based on sequencing read length. It takes as input
a pheno table of all the same rls and queries the bigwig_path.
}
\keyword{getExCounts}
