% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRseTx.R
\name{getRseTx}
\alias{getRseTx}
\title{Download and/or access quantified SRA projects}
\usage{
getRseTx(project, download_path = NULL)
}
\arguments{
\item{project}{A SRA project name for a project already curated in recount2.}

\item{download_path}{Where the rse should be downloaded to if not NULL.
Defaults to NULL, i.e. no downloading.}
}
\value{
Returns the path to downloaded object.
}
\description{
This function downloads and/or loads into R a
\code{rangedSummarizedExperiment} object for transcript level abundance
produced by the latest version of recountNNLS.
}
\examples{
project = 'DRP000366'
getRseTx(project)
}
\keyword{recountNNLS}
