% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPheno.R
\name{processPheno}
\alias{processPheno}
\title{Phenotype Information Creation}
\usage{
processPheno(input, local = F)
}
\arguments{
\item{input}{The project name from SRA or a manifest file.}
}
\description{
This function creates a phenotype matrix with the information needed
for the rest of the analysis. The supplied input is either an SRA
project id (from a project compiled on recount2), or a manifest file
that contains information where the rail outputs are.
}
\keyword{processPheno}
