% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPheno.R
\name{processPheno}
\alias{processPheno}
\title{Phenotype Information Creation}
\usage{
processPheno(input)
}
\arguments{
\item{input}{The project name from SRA.}
}
\description{
This function creates a phenotype matrix with the information needed
to be fed into recountNNLS().
}
\keyword{processPheno}
