% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processReadLength.R
\name{processReadLength}
\alias{processReadLength}
\title{NNLS for Tx Abundance Calculation}
\usage{
processReadLength(rl, pheno, counts_ex, counts_jx, cores)
}
\arguments{
\item{rl}{A read length that exists in the pheno table}

\item{pheno}{The phenotype matrix created by processPheno().}

\item{cores}{The number of processing cores to use.}
}
\description{
This function applies NNLS to calculate the Tx abundance
from coverage of exonic and junction coverage statistics extracted using getExCounts()
and getJxCounts().
}
\keyword{recountNNLS}
