% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processReadLength.R
\name{processReadLength}
\alias{processReadLength}
\title{NNLS for Tx Abundance Calculation}
\usage{
processReadLength(rl, pheno, jx_file, cores, power = 0)
}
\arguments{
\item{rl}{A read length that exists in the pheno table}

\item{pheno}{The phenotype matrix created by processPheno().}

\item{jx_file}{The path to the Rail-RNA junction coverage file if processing sample not already in recount2.}

\item{cores}{The number of processing cores to use.}
}
\value{
A rse of the quantified tx abundances in the samples in pheno.
}
\description{
This function applies runs our NNLS model for a given subset
of the pheno matrix that has the same read length specified
and reports an rse.
}
\keyword{recountNNLS}
