% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recountNNLS.R
\name{recountNNLS}
\alias{recountNNLS}
\title{NNLS for Tx Abundance Calculation}
\usage{
recountNNLS(pheno, counts_ex, counts_jx, cores = 1)
}
\arguments{
\item{pheno}{The table of phenotype information from processPheno().}

\item{cores}{The number of processing cores to use.}
}
\description{
This function applies NNLS to calculate the Tx abundance
from coverage of exonic and junction coverage statistics extracfted using getExCounts()
and getJxCounts().
}
\keyword{recountNNLS}
