% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recountNNLS.R
\name{recountNNLS}
\alias{recountNNLS}
\title{NNLS for Tx Abundance Calculation}
\usage{
recountNNLS(pheno, cores = 1)
}
\arguments{
\item{pheno}{The a table of phenotype information from processPheno().}

\item{cores}{The number of processing cores to use.}
}
\description{
This function will use a NNLS to calculate the Tx abundance
using reduced-representation of coverage information
extracted from rail output available on recount2
}
\keyword{recountNNLS}
